api/requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
Flask-RESTful==0.3.9
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
urllib3==1.26.9
waitress==2.1.1
Werkzeug==2.1.1
